--- proof/proof.py
@@ -1,2 +1,0 @@
-cuadrados = [x**2 for x in range(10)]
-print(cuadrados)
