+++ src/start_todo/todo_manager.py
@@ -1,5 +1,5 @@
 #todo_manager.py
-from .todo import Todo
+from .models.todo import Todo
 import json
 class TodoManager:
     def __init__(self):
@@ -28,18 +28,21 @@ class TodoManager:
     def clear_todos(self):
         self.todos = []
     def save_to_file(self):
-        with open('./Todos.json', 'w') as file:
+        with open('./src/data/Todos.json', 'w') as file:
             json.dump(self.todos, file, default=lambda obj: obj.__dict__, indent=2)
     def show_tasks(self):
         for todo in self.todos:
             print(todo.__str__())
     def read_to_file(self):
-        with open('./Todos.json', 'r') as file:
-            todo_data= json.load(file)
-            if(todo_data == None):
-                self.save_to_file()
-            else:
+        try:
+            with open('./src/data/Todos.json', 'r') as file:
+                todo_data = json.load(file)
                 for todo in todo_data:
                     task = Todo(todo.get('title', ''), todo.get('description', ''), todo.get('category',''), todo.get('state',bool))
                     self.todos.append(task)
+                return True
+        except FileNotFoundError:
+            print("File not found")
+            return None
+
