+++ src/todo_begin.py
@@ -1,0 +1,6 @@
+#!/usr/bin/env python3
+#import sys
+from start_todo.todo_promp import TodoProm
+
+if __name__ == "__main__":
+    TodoProm.todo_promp()
